+++ MagicVilla_API/Controllers/v1/NumeroVillaController.cs
@@ -10,21 +10,25 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Net;
 
-namespace MagicVilla_API.Controllers
+namespace MagicVilla_API.Controllers.v1
 {
-    [Route("api/[controller]")]
+    [Route("api/v{version:apiVersion}/[controller]")]
     [ApiController]
-    public class VillaController : ControllerBase
+    [ApiVersion("1.0")]
+    public class NumeroVillaController : ControllerBase
     {
-        private readonly ILogger<VillaController> _logger;
+        private readonly ILogger<NumeroVillaController> _logger;
         private readonly IVillaRepositorio _villaRepo;
+        private readonly INumeroVillaRepositorio _numeroRepo;
         private readonly IMapper _mapper;
         protected APIResponse _response;
 
-        public VillaController(ILogger<VillaController> logger, IVillaRepositorio villaRepo, IMapper mapper)
+        public NumeroVillaController(ILogger<NumeroVillaController> logger,
+                    IVillaRepositorio villaRepo, INumeroVillaRepositorio numeroRepo, IMapper mapper)
         {
             _logger = logger;
             _villaRepo = villaRepo;
+            _numeroRepo = numeroRepo;
             _mapper = mapper;
             _response = new();
         }
@@ -32,14 +36,14 @@ namespace MagicVilla_API.Controllers
         [HttpGet]
         [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult<APIResponse>> GetVillas() 
+        public async Task<ActionResult<APIResponse>> GetNumeroVillas()
         {
             try
             {
                 _logger.LogInformation("Obtener todas la Villas");
 
-                IEnumerable<Villa> villaList = await _villaRepo.ObtenerTodos();
-                _response.Resultado = _mapper.Map<IEnumerable<VillaDto>>(villaList);
+                IEnumerable<NumeroVilla> numerovillaList = await _numeroRepo.ObtenerTodos(incluirPropiedades: "Villa");
+                _response.Resultado = _mapper.Map<IEnumerable<NumeroVillaDto>>(numerovillaList);
                 _response.statusCode = HttpStatusCode.OK;
 
                 return Ok(_response);
@@ -53,33 +57,34 @@ namespace MagicVilla_API.Controllers
             return _response;
         }
 
-        [HttpGet("{id:int}", Name ="GetVilla")]
+
+        [HttpGet("{id:int}", Name = "GetNumeroVilla")]
         [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<ActionResult<APIResponse>> GetVilla(int id) 
+        public async Task<ActionResult<APIResponse>> GetNumeroVilla(int id)
         {
             try
             {
                 if (id == 0)
                 {
-                    _logger.LogError("Error al traer la villa con el id " + id);
+                    _logger.LogError("Error al traer el numero de villa con el id " + id);
                     _response.IsExitoso = false;
                     _response.statusCode = HttpStatusCode.BadRequest;
                     return BadRequest(_response);
                 }
 
-                var villa = await _villaRepo.Obtener(v => v.Id == id);
-                if (villa == null)
+                var numerovilla = await _numeroRepo.Obtener(v => v.VillaNo == id, incluirPropiedades: "Villa");
+                if (numerovilla == null)
                 {
-                    _logger.LogError("Error al traer la villa con el id " + id);
+                    _logger.LogError("Error al traer el numero de villa con el id " + id);
                     _response.IsExitoso = false;
                     _response.statusCode = HttpStatusCode.NotFound;
                     return NotFound(_response);
                 }
 
-                _response.Resultado = _mapper.Map<VillaDto>(villa);
+                _response.Resultado = _mapper.Map<NumeroVillaDto>(numerovilla);
                 _response.statusCode = HttpStatusCode.OK;
                 return Ok(_response);
             }
@@ -94,11 +99,11 @@ namespace MagicVilla_API.Controllers
         }
 
         [HttpPost]
-        [Authorize(Roles ="admin")]
+        [Authorize(Roles = "admin")]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<APIResponse>> CrearVilla([FromBody] VillaCreateDto createDto)
+        public async Task<ActionResult<APIResponse>> CrearNumeroVilla([FromBody] NumeroVillaCreateDto createDto)
         {
             try
             {
@@ -107,10 +112,15 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(ModelState);
                 }
 
-                if (await _villaRepo.Obtener(v => v.Nombre.ToLower()
-                        == createDto.Nombre.ToLower()) != null)
+                if (await _numeroRepo.Obtener(v => v.VillaNo == createDto.VillaNo) != null)
                 {
-                    ModelState.AddModelError("ErrorMessages", "Ya existe una billa con ese Nombre");
+                    ModelState.AddModelError("ErrorMessages", "El numero de villa ya Existe");
+                    return BadRequest(ModelState);
+                }
+
+                if (await _villaRepo.Obtener(v => v.Id == createDto.VillaId) == null)
+                {
+                    ModelState.AddModelError("ErrorMessages", "El Id de la villa no existe");
                     return BadRequest(ModelState);
                 }
 
@@ -120,15 +130,15 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(createDto);
                 }
 
-                Villa modelo = _mapper.Map<Villa>(createDto);
+                NumeroVilla modelo = _mapper.Map<NumeroVilla>(createDto);
 
                 modelo.FechaCreacion = DateTime.Now;
                 modelo.FechaActualizacion = DateTime.Now;
-                await _villaRepo.Crear(modelo);
+                await _numeroRepo.Crear(modelo);
                 _response.Resultado = modelo;
                 _response.statusCode = HttpStatusCode.Created;
 
-                return CreatedAtRoute("GetVilla", new { id = modelo.Id }, _response);
+                return CreatedAtRoute("GetVilla", new { id = modelo.VillaNo }, _response);
             }
             catch (Exception ex)
             {
@@ -136,7 +146,7 @@ namespace MagicVilla_API.Controllers
                 _response.IsExitoso = false;
                 _response.ErrorMensages = new List<string>() { ex.ToString() };
             }
-            return _response;          
+            return _response;
         }
 
         [HttpDelete("{id:int}")]
@@ -144,7 +154,7 @@ namespace MagicVilla_API.Controllers
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> DeleteVilla(int id)
+        public async Task<IActionResult> DeleteNumeroVilla(int id)
         {
             try
             {
@@ -155,15 +165,15 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(_response);
                 }
 
-                var villa = await _villaRepo.Obtener(v => v.Id == id);
-                if (villa == null)
+                var numerovilla = await _numeroRepo.Obtener(v => v.VillaNo == id);
+                if (numerovilla == null)
                 {
                     _response.IsExitoso = false;
                     _response.statusCode = HttpStatusCode.NotFound;
                     return NotFound(_response);
                 }
 
-                await _villaRepo.Remover(villa);
+                await _numeroRepo.Remover(numerovilla);
 
                 _response.statusCode = HttpStatusCode.NoContent;
 
@@ -175,18 +185,18 @@ namespace MagicVilla_API.Controllers
                 _response.IsExitoso = false;
                 _response.ErrorMensages = new List<string>() { ex.ToString() };
             }
-            return BadRequest(_response);           
+            return BadRequest(_response);
         }
 
         [HttpPut("{id:int}")]
         [Authorize(Roles = "admin")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> UpdateVilla(int id, [FromBody] VillaUpdateDto updateDto)
+        public async Task<IActionResult> UpdateNumeroVilla(int id, [FromBody] NumeroVillaUpdateDto updateDto)
         {
             try
             {
-                if (updateDto == null || id != updateDto.Id)
+                if (updateDto == null || id != updateDto.VillaNo)
                 {
                     _logger.LogError("Error al traer la villa con el id " + id);
                     _response.IsExitoso = false;
@@ -194,10 +204,16 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(_response);
                 }
 
-                Villa modelo = _mapper.Map<Villa>(updateDto);
+                if (await _villaRepo.Obtener(v => v.Id == updateDto.VillaId) == null)
+                {
+                    ModelState.AddModelError("ErrorMessages", "El Id de la villa no existe");
+                    return BadRequest(ModelState);
+                }
 
-                modelo.FechaActualizacion = DateTime.Now;
-                await _villaRepo.Actualizar(modelo);
+                NumeroVilla modelo = _mapper.Map<NumeroVilla>(updateDto);
+
+
+                await _numeroRepo.Actualizar(modelo);
                 _response.statusCode = HttpStatusCode.NoContent;
 
                 return Ok(_response);
@@ -207,42 +223,8 @@ namespace MagicVilla_API.Controllers
                 _response.IsExitoso = false;
                 _response.ErrorMensages = new List<string>() { ex.ToString() };
             }
-            return Ok(_response);            
-        }
-
-        [HttpPatch("{id:int}")]
-        [Authorize(Roles = "admin")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> UpdatePacialVilla(int id, JsonPatchDocument<VillaUpdateDto> patchDto)
-        {
-            if (patchDto == null || id == 0)
-            {
-                _logger.LogError("Error al traer la villa con el id " + id);
-                _response.IsExitoso = false;
-                _response.statusCode = HttpStatusCode.BadRequest;
-                return BadRequest(_response);
-            }
-
-            var villa = await _villaRepo.Obtener(v => v.Id == id, tracked:false);
-
-            VillaUpdateDto villaDto = _mapper.Map<VillaUpdateDto>(villa);
-
-            if (villa == null) return BadRequest();
-
-            patchDto.ApplyTo(villaDto, ModelState);
-
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            Villa modelo = _mapper.Map<Villa>(villaDto);
-
-            await _villaRepo.Actualizar(modelo);
-            _response.statusCode = HttpStatusCode.NoContent;
-
             return Ok(_response);
         }
+
     }
 }
