+++ MagicVilla_API/Controllers/v1/VillaController.cs
@@ -10,24 +10,22 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Net;
 
-namespace MagicVilla_API.Controllers
+namespace MagicVilla_API.Controllers.v1
 {
-    [Route("api/[controller]")]
+    [Route("api/v{version:apiVersion}/[controller]")]
     [ApiController]
-    public class NumeroVillaController : ControllerBase
+    [ApiVersion("1.0")]
+    public class VillaController : ControllerBase
     {
-        private readonly ILogger<NumeroVillaController> _logger;
+        private readonly ILogger<VillaController> _logger;
         private readonly IVillaRepositorio _villaRepo;
-        private readonly INumeroVillaRepositorio _numeroRepo;
         private readonly IMapper _mapper;
         protected APIResponse _response;
 
-        public NumeroVillaController(ILogger<NumeroVillaController> logger, 
-                    IVillaRepositorio villaRepo,INumeroVillaRepositorio numeroRepo ,IMapper mapper)
+        public VillaController(ILogger<VillaController> logger, IVillaRepositorio villaRepo, IMapper mapper)
         {
             _logger = logger;
             _villaRepo = villaRepo;
-            _numeroRepo = numeroRepo;
             _mapper = mapper;
             _response = new();
         }
@@ -35,14 +33,14 @@ namespace MagicVilla_API.Controllers
         [HttpGet]
         [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult<APIResponse>> GetNumeroVillas() 
+        public async Task<ActionResult<APIResponse>> GetVillas()
         {
             try
             {
                 _logger.LogInformation("Obtener todas la Villas");
 
-                IEnumerable<NumeroVilla> numerovillaList = await _numeroRepo.ObtenerTodos(incluirPropiedades:"Villa");
-                _response.Resultado = _mapper.Map<IEnumerable<NumeroVillaDto>>(numerovillaList);
+                IEnumerable<Villa> villaList = await _villaRepo.ObtenerTodos();
+                _response.Resultado = _mapper.Map<IEnumerable<VillaDto>>(villaList);
                 _response.statusCode = HttpStatusCode.OK;
 
                 return Ok(_response);
@@ -56,33 +54,33 @@ namespace MagicVilla_API.Controllers
             return _response;
         }
 
-        [HttpGet("{id:int}", Name ="GetNumeroVilla")]
+        [HttpGet("{id:int}", Name = "GetVilla")]
         [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<ActionResult<APIResponse>> GetNumeroVilla(int id) 
+        public async Task<ActionResult<APIResponse>> GetVilla(int id)
         {
             try
             {
                 if (id == 0)
                 {
-                    _logger.LogError("Error al traer el numero de villa con el id " + id);
+                    _logger.LogError("Error al traer la villa con el id " + id);
                     _response.IsExitoso = false;
                     _response.statusCode = HttpStatusCode.BadRequest;
                     return BadRequest(_response);
                 }
 
-                var numerovilla = await _numeroRepo.Obtener(v => v.VillaNo == id, incluirPropiedades:"Villa");
-                if (numerovilla == null)
+                var villa = await _villaRepo.Obtener(v => v.Id == id);
+                if (villa == null)
                 {
-                    _logger.LogError("Error al traer el numero de villa con el id " + id);
+                    _logger.LogError("Error al traer la villa con el id " + id);
                     _response.IsExitoso = false;
                     _response.statusCode = HttpStatusCode.NotFound;
                     return NotFound(_response);
                 }
 
-                _response.Resultado = _mapper.Map<NumeroVillaDto>(numerovilla);
+                _response.Resultado = _mapper.Map<VillaDto>(villa);
                 _response.statusCode = HttpStatusCode.OK;
                 return Ok(_response);
             }
@@ -101,7 +99,7 @@ namespace MagicVilla_API.Controllers
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<APIResponse>> CrearNumeroVilla([FromBody] NumeroVillaCreateDto createDto)
+        public async Task<ActionResult<APIResponse>> CrearVilla([FromBody] VillaCreateDto createDto)
         {
             try
             {
@@ -110,15 +108,10 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(ModelState);
                 }
 
-                if (await _numeroRepo.Obtener(v => v.VillaNo == createDto.VillaNo) != null)
+                if (await _villaRepo.Obtener(v => v.Nombre.ToLower()
+                        == createDto.Nombre.ToLower()) != null)
                 {
-                    ModelState.AddModelError("ErrorMessages", "El numero de villa ya Existe");
-                    return BadRequest(ModelState);
-                }
-
-                if(await _villaRepo.Obtener(v=>v.Id==createDto.VillaId) == null)
-                {
-                    ModelState.AddModelError("ErrorMessages", "El Id de la villa no existe");
+                    ModelState.AddModelError("ErrorMessages", "Ya existe una billa con ese Nombre");
                     return BadRequest(ModelState);
                 }
 
@@ -128,15 +121,15 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(createDto);
                 }
 
-                NumeroVilla modelo = _mapper.Map<NumeroVilla>(createDto);
+                Villa modelo = _mapper.Map<Villa>(createDto);
 
                 modelo.FechaCreacion = DateTime.Now;
                 modelo.FechaActualizacion = DateTime.Now;
-                await _numeroRepo.Crear(modelo);
+                await _villaRepo.Crear(modelo);
                 _response.Resultado = modelo;
                 _response.statusCode = HttpStatusCode.Created;
 
-                return CreatedAtRoute("GetVilla", new { id = modelo.VillaNo }, _response);
+                return CreatedAtRoute("GetVilla", new { id = modelo.Id }, _response);
             }
             catch (Exception ex)
             {
@@ -144,7 +137,7 @@ namespace MagicVilla_API.Controllers
                 _response.IsExitoso = false;
                 _response.ErrorMensages = new List<string>() { ex.ToString() };
             }
-            return _response;          
+            return _response;
         }
 
         [HttpDelete("{id:int}")]
@@ -152,7 +145,7 @@ namespace MagicVilla_API.Controllers
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> DeleteNumeroVilla(int id)
+        public async Task<IActionResult> DeleteVilla(int id)
         {
             try
             {
@@ -163,15 +156,15 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(_response);
                 }
 
-                var numerovilla = await _numeroRepo.Obtener(v => v.VillaNo == id);
-                if (numerovilla == null)
+                var villa = await _villaRepo.Obtener(v => v.Id == id);
+                if (villa == null)
                 {
                     _response.IsExitoso = false;
                     _response.statusCode = HttpStatusCode.NotFound;
                     return NotFound(_response);
                 }
 
-                await _numeroRepo.Remover(numerovilla);
+                await _villaRepo.Remover(villa);
 
                 _response.statusCode = HttpStatusCode.NoContent;
 
@@ -183,18 +176,18 @@ namespace MagicVilla_API.Controllers
                 _response.IsExitoso = false;
                 _response.ErrorMensages = new List<string>() { ex.ToString() };
             }
-            return BadRequest(_response);           
+            return BadRequest(_response);
         }
 
         [HttpPut("{id:int}")]
         [Authorize(Roles = "admin")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> UpdateNumeroVilla(int id, [FromBody] NumeroVillaUpdateDto updateDto)
+        public async Task<IActionResult> UpdateVilla(int id, [FromBody] VillaUpdateDto updateDto)
         {
             try
             {
-                if (updateDto == null || id != updateDto.VillaNo)
+                if (updateDto == null || id != updateDto.Id)
                 {
                     _logger.LogError("Error al traer la villa con el id " + id);
                     _response.IsExitoso = false;
@@ -202,16 +195,10 @@ namespace MagicVilla_API.Controllers
                     return BadRequest(_response);
                 }
 
-                if(await _villaRepo.Obtener(v=>v.Id == updateDto.VillaId) == null)
-                {
-                    ModelState.AddModelError("ErrorMessages", "El Id de la villa no existe");
-                    return BadRequest(ModelState);
-                }
+                Villa modelo = _mapper.Map<Villa>(updateDto);
 
-                NumeroVilla modelo = _mapper.Map<NumeroVilla>(updateDto);
-
-
-                await _numeroRepo.Actualizar(modelo);
+                modelo.FechaActualizacion = DateTime.Now;
+                await _villaRepo.Actualizar(modelo);
                 _response.statusCode = HttpStatusCode.NoContent;
 
                 return Ok(_response);
@@ -221,8 +208,42 @@ namespace MagicVilla_API.Controllers
                 _response.IsExitoso = false;
                 _response.ErrorMensages = new List<string>() { ex.ToString() };
             }
-            return Ok(_response);            
+            return Ok(_response);
         }
-                
+
+        [HttpPatch("{id:int}")]
+        [Authorize(Roles = "admin")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> UpdatePacialVilla(int id, JsonPatchDocument<VillaUpdateDto> patchDto)
+        {
+            if (patchDto == null || id == 0)
+            {
+                _logger.LogError("Error al traer la villa con el id " + id);
+                _response.IsExitoso = false;
+                _response.statusCode = HttpStatusCode.BadRequest;
+                return BadRequest(_response);
+            }
+
+            var villa = await _villaRepo.Obtener(v => v.Id == id, tracked: false);
+
+            VillaUpdateDto villaDto = _mapper.Map<VillaUpdateDto>(villa);
+
+            if (villa == null) return BadRequest();
+
+            patchDto.ApplyTo(villaDto, ModelState);
+
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            Villa modelo = _mapper.Map<Villa>(villaDto);
+
+            await _villaRepo.Actualizar(modelo);
+            _response.statusCode = HttpStatusCode.NoContent;
+
+            return Ok(_response);
+        }
     }
 }
