+++ MagicVilla_API/Controllers/v2/NumeroVillaController.cs
@@ -1,0 +1,42 @@
+﻿using AutoMapper;
+using MagicVilla_API.Datos;
+using MagicVilla_API.Modelos;
+using MagicVilla_API.Modelos.Dto;
+using MagicVilla_API.Repositorio.IRepositorio;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.JsonPatch;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System.Net;
+
+namespace MagicVilla_API.Controllers.v2
+{
+    [Route("api/v{version:apiVersion}/[controller]")]
+    [ApiController]
+    [ApiVersion("2.0")]
+    public class NumeroVillaController : ControllerBase
+    {
+        private readonly ILogger<NumeroVillaController> _logger;
+        private readonly IVillaRepositorio _villaRepo;
+        private readonly INumeroVillaRepositorio _numeroRepo;
+        private readonly IMapper _mapper;
+        protected APIResponse _response;
+
+        public NumeroVillaController(ILogger<NumeroVillaController> logger, IVillaRepositorio villaRepo,
+                                                                            INumeroVillaRepositorio numeroRepo, IMapper mapper)
+        {
+            _logger = logger;
+            _villaRepo = villaRepo;
+            _numeroRepo = numeroRepo;
+            _mapper = mapper;
+            _response = new();
+        }
+
+        [HttpGet]
+        public IEnumerable<string> Get()
+        {
+            return new string[] { "valor1", "valor2" };
+        }
+
+    }
+}
