+++ MagicVilla_Web/Services/BaseService.cs
@@ -66,7 +66,8 @@ namespace MagicVilla_Web.Services
                 try
                 {
                     APIResponse response = JsonConvert.DeserializeObject<APIResponse>(apiContent);
-                    if (apiResponse.StatusCode == HttpStatusCode.BadRequest || apiResponse.StatusCode == HttpStatusCode.NotFound)
+                    if (response != null && (apiResponse.StatusCode == HttpStatusCode.BadRequest 
+                        || apiResponse.StatusCode == HttpStatusCode.NotFound))
                     {
                         response.statusCode = HttpStatusCode.BadRequest;
                         response.IsExitoso = false;
