+++ MagicVilla_Web/Services/UsuarioService.cs
@@ -22,7 +22,7 @@ namespace MagicVilla_Web.Services
             {
                 APITipo = DS.APITipo.POST,
                 Datos = dto,
-                Url = _villaUrl + "/api/usuario/login"
+                Url = _villaUrl + "/api/v1/usuario/login"
             });
         }
 
@@ -32,7 +32,7 @@ namespace MagicVilla_Web.Services
             {
                 APITipo = DS.APITipo.POST,
                 Datos = dto,
-                Url = _villaUrl + "/api/usuario/registrar"
+                Url = _villaUrl + "/api/v1/usuario/registrar"
             });
         }
     }
