+++ MagicVilla_Web/Services/VillaService.cs
@@ -23,7 +23,7 @@ namespace MagicVilla_Web.Services
             {
                 APITipo = DS.APITipo.PUT,
                 Datos = dto,
-                Url = _villaUrl + "/api/Villa/"+dto.Id,
+                Url = _villaUrl + "/api/v1/Villa/"+dto.Id,
                 Token = token
             });
         }
@@ -34,7 +34,7 @@ namespace MagicVilla_Web.Services
             {
                 APITipo = DS.APITipo.POST,
                 Datos = dto,
-                Url = _villaUrl+"/api/Villa",
+                Url = _villaUrl+"/api/v1/Villa",
                 Token = token
             });
         }
@@ -44,7 +44,7 @@ namespace MagicVilla_Web.Services
             return SendAsync<T>(new APIRequest()
             {
                 APITipo = DS.APITipo.GET,
-                Url = _villaUrl + "/api/Villa/" + id,
+                Url = _villaUrl + "/api/v1/Villa/" + id,
                 Token = token
             });
         }
@@ -54,7 +54,7 @@ namespace MagicVilla_Web.Services
             return SendAsync<T>(new APIRequest()
             {
                 APITipo = DS.APITipo.GET,
-                Url = _villaUrl + "/api/Villa",
+                Url = _villaUrl + "/api/v1/Villa",
                 Token = token
             });
         }
@@ -64,7 +64,7 @@ namespace MagicVilla_Web.Services
             return SendAsync<T>(new APIRequest()
             {
                 APITipo = DS.APITipo.DELETE,
-                Url = _villaUrl + "/api/Villa/" + id,
+                Url = _villaUrl + "/api/v1/Villa/" + id,
                 Token = token
             });
         }
